Introduce hashbrown test, failing
On a tip from @Amanieu I introduced a test of the hashbrown Rust library.
This test fails on account of we're still kicking out invalid memory
alignments, looks like. On the plus side this integration test paves
the way for getting rid of the bin targets and focusing entirely on
CI.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>