Transition bins into integration tests, up byte alignments
This commit bumps the minimum byte alignment to 8 bytes on 32 bit
platforms and 16 bytes on 64 bit platforms, per the docs I could
find in glibc and on Eric Rahm's discussion of the Firefox allocation
subsystem. The hashbrown suite still fails.

I've also found that having such a large static section is a real
pain for compile times. Investing in proper dynamic allocation is
almost surely going to be the thing to do here shortly.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>