Introduce dynamic allocation
Previously the bump allocator used a constant, static array as an
allocated pool. This was nice in the sense that it was very easy to
produce but had a major downside in the link size of any executable
using bh_alloc. This is corrected by using mmap to get some heap space
on the first allocation of the system.

bh_alloc using programs should now be much smaller.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>